+++ server/controller/routes.js
@@ -2,6 +2,10 @@
 import clientRequest from '../lib/client';
 
 module.exports = {
+  getImport(req, res) {
+    const importedData = clientRequest.getImport();
+    res.status(200).json(importedData);
+  },
   getAll(req, res) {
     const allOrganisation = clientRequest.getAllOrganisation();
     allOrganisation.then(orgData => res.status(200).json({ data: orgData }));
@@ -11,7 +15,8 @@ module.exports = {
     catagories.then(category => res.status(200).json({ data: category }));
   },
   getOrganisation(req, res) {
-    const categoryName = req.body;
-    const organisation = clientRequest.getOrganisation(categoryName);
+    const categoryName = req.params.category;
+    const organisations = clientRequest.getOrganisation(categoryName);
+    organisations.then(organisation => res.status(200).json({ data: organisation }));
   },
 };
